separate VirtualNode.js
