+++ src/createContainerElement.js
@@ -1,0 +1,6 @@
+function createContainerElement() {
+  const el = document.createElement('div');
+  el.className = 'json-container';
+  return el;
+}
+export {createContainerElement};
