+++ src/index.js
@@ -2,3 +2,4 @@ export * from './getDataType.js';
 export * from './json-view.js';
 export * from './listen.js';
 export * from './traverseObject.js';
+export * from './VirtualNode.js';
